requirements.txt
fastapi
pymongo
sentence-transformers
uvicorn
python-dotenv
numpy
python-multipart
python-dotenv